requirements.txt
psycopg2==2.9.4
Django==4.1.2
Pillow==9.2.0
django-durationwidget==1.0.5
djangorestframework==3.14.0